Remove config
The contents of the config were enabling/disabling a feature. However, it was a "soft" disable, as in the block only became unobtainable through crafting or creative inventory and prevented Golem creation. The reason for removal is basically a redundancy: datapacks and modpack tools can enable or disable these features.

As these were the only options, the config has been removed. Should any config in the future exist such as builder overrides, it will be returned.